requirements.txt
-r RAG/requirements.txt
numpy >= 2.3.3
torchvision==0.26.0
torch==2.11.0
beautifulsoup4==4.14.3
chromadb==1.5.8
Flask==3.1.3
mss==10.1.0
Pillow==12.2.0
pydantic_settings==2.13.1
Requests==2.33.1
transformers==5.6.2
bitsandbytes @ https://github.com/bitsandbytes-foundation/bitsandbytes/releases/download/continuous-release_main/bitsandbytes-1.33.7.preview-py3-none-win_amd64.whl
turboquant==0.2.0 
accelerate==1.6.0

llama-cpp-python==0.3.21